Update pandas_practice.py
"2. Data Manipulation

a. Sort the DataFrame based on the Age column in descending order.

b. Create a new column called Senior which marks True if the age is above 60, and False otherwise.

c. Replace a city name of your choice with another city name for all occurrences."

Done. Took less than an hour of searching and debugging code.